Initial optional tip support for trigger event delegation...
Provided `$triggerContext` is set, we can get
enough data for an event delegation call. The main
change is _bindTrigger, which now delegates if the
given $trigger is omitted. Note that a false-y
selector argument is interpreted as direct
binding. Additionally, the event extension's
binding calls need to support delegation. The
delegation selector has to be passed through to
the special event's setup method, which is
basically wrapping more event bindings, and the
best bridge for now seems to be the event
binding-to-handler data argument.

_bindTrigger has also been simplified into from two base handlers to one.