Rework tip plugin's context support...
The combo of triggerContext and context has now
been replaced with the more understandable context
and viewport elements, respectively. The viewport
option defaults to body, and the context should be
passed in much like triggerContext before, except
as a plugin argument, and it will also default to
body.

This fixes the issue in the visual test where the
final call wasn't going through.