Reimplement media-grid dimming and focusing.
Correctly dim and un-dim media-grid on hover. _handleEnteringDimmedItem
is simpler as handling media-grid mouseleave.

Dimming is part of focusing, which can be grouped into toggleItemFocus.
This means renaming toggleItem and wrapping a toggleExpandedItemFocus,
as well as renaming dimDelay.

Also correctly resets $expandedItem. Also adds a 1s delay of focus when
leaving expanded item that can be cancelled via _clearDimTimeout on
expanding another item.